+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="5.0.1"></a>
+## 5.0.1 (2016-04-22)
+
+
+### Features
+
+* **Dropdown:** Support passing your own toggle onClick handler ([60bf3a9](https://github.com/pivotal-cf/pivotal-ui/commit/60bf3a9))
+* **DropdownItem:** Pass additional props if DropdownItem has an href ([d0afeee](https://github.com/pivotal-cf/pivotal-ui/commit/d0afeee))
+* **Radio:** render input before label for more customizable styling ([a969ea3](https://github.com/pivotal-cf/pivotal-ui/commit/a969ea3))
+
+
 <a name="5.0.0"></a>
 # 5.0.0 (2016-04-18)